+++ conductivity_msg.h
